Feat(Signout): ユーザがログインしているときにサインアウトボタンを表示する
